app/requirements.txt
cffi==1.11.2
falcon==1.4.1
gunicorn==19.7.1
python-mimeparse==1.6.0
readline==6.2.4.1
six==1.11.0
gitpython==3.0.5